test: split RpcTests into e2e (server) + RpcUnitTests (mock) suites
Pulls the 12 mock-based tests (caller-side state machine, server-side handler
dispatch, v2 wire-format / state-machine internals) out of `RpcTests` into a
new `RpcUnitTests` suite. The new suite is untagged and runs parallel by
default; `RpcTests` keeps `.serialized .tags(.e2e)` and only contains the 7
real-WebRTC tests that go through `withRooms([...])`.

Benefits:
- `.tags(.e2e)` now accurately gates suite execution by "needs a server".
- Unit tests run in parallel (Swift Testing default), shrinking CI time.
- `RpcTests.swift` drops from 784 → 210 lines and no longer needs the
  `swiftlint:disable file_length type_body_length` block. `RpcUnitTests.swift`
  keeps the disables; the body is wide by design (15 tests across 3 sub-areas).

Test support helpers (`RpcTestSupport`, `TestStringCollector`) move with their
only consumers into `RpcUnitTests.swift`. No behavioural change to any test.
